# resources/openapi/user-api.yaml
openapi: 3.1.0
info:
    title: 'fvn.li User API'
    version: 1.0.0
    license:
        name: MIT
        identifier: MIT
    description: "Visible catalog games and the token owner's private tags, lists, tracking, and search preferences. Ratings, reviews, profile, export, account deletion, social account linking, and token management are website-only."
servers:
    - url: /api/v1
security:
    - bearerAuth: []
...
paths:
    /games:
        get:
            tags:
                - Catalog
            operationId: getGames
            summary: 'Search visible catalog games'
            x-required-ability: 'catalog:read'
            parameters:
                - name: q
                  in: query
                  schema:
                      type: string
                      maxLength: 100
                - name: platform
                  in: query
                  schema:
                      type: string
                      enum:
                          - itch_io
                          - steam
                          - other
                - name: per_page
                  in: query
                  schema:
                      type: integer
                      minimum: 1
                      maximum: 50
                - $ref: '#/components/parameters/Page'
            responses:
                '200':
                    description: Games
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GamePage'
                            example:
                                data:
                                    - id: 368
                                      name: 'TRAINEE DETECTIVE (EN)'
                                      slug: trainee-detective-en
                                      platform: itch_io
                                      itch_id: 2241851
                                      steam_app_id: null
                                      urls:
                                          itch_io: 'https://starnathy-arts.itch.io/trainee-detective-en'
                                      status: Released
                                      is_nsfw: false
                                      is_paid: true
                                      has_demo: true
                                      fvn_url: 'https://fvn.li/games/trainee-detective-en'
                                    - id: 402963
                                      name: 'Dragon Detective: A Friend from the Shadows'
                                      slug: dragon-detective-a-friend-from-the-shadows
                                      platform: steam
                                      itch_id: null
                                      steam_app_id: 3640850
                                      urls:
                                          steam: 'https://store.steampowered.com/app/3640850/Dragon_Detective_A_Friend_from_the_Shadows/'
                                      status: Released
                                      is_nsfw: false
                                      is_paid: true
                                      has_demo: true
                                      fvn_url: 'https://fvn.li/games/dragon-detective-a-friend-from-the-shadows'
                                links:
                                    first: 'https://fvn.li/api/v1/games?q=detective&per_page=2&page=1'
                                    last: 'https://fvn.li/api/v1/games?q=detective&per_page=2&page=2'
                                    prev: null
                                    next: 'https://fvn.li/api/v1/games?q=detective&per_page=2&page=2'
                                meta:
                                    current_page: 1
                                    from: 1
                                    last_page: 2
                                    links:
                                        - url: null
                                          label: '&laquo; Previous'
                                          page: null
                                          active: false
                                        - url: 'https://fvn.li/api/v1/games?q=detective&per_page=2&page=1'
                                          label: '1'
                                          page: 1
                                          active: true
                                        - url: 'https://fvn.li/api/v1/games?q=detective&per_page=2&page=2'
                                          label: '2'
                                          page: 2
                                          active: false
                                        - url: 'https://fvn.li/api/v1/games?q=detective&per_page=2&page=2'
                                          label: 'Next &raquo;'
                                          page: 2
                                          active: false
                                    path: 'https://fvn.li/api/v1/games'
                                    per_page: 2
                                    to: 2
                                    total: 3
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '422':
                    $ref: '#/components/responses/ValidationError'
                '429':
                    $ref: '#/components/responses/RateLimited'
...
    '/games/{game}':
        get:
            tags:
                - Catalog
            operationId: getGamesGame
            summary: 'Get a visible catalog game'
            x-required-ability: 'catalog:read'
            parameters:
                - $ref: '#/components/parameters/Game'
            responses:
                '200':
                    description: Game
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GameResponse'
                            example:
                                data:
                                    id: 407746
                                    name: 'Chicken Police - Paint it RED!'
                                    slug: chicken-police-paint-it-red
                                    platform: steam
                                    itch_id: null
                                    steam_app_id: 1084640
                                    urls:
                                        steam: 'https://store.steampowered.com/app/1084640/Chicken_Police__Paint_it_RED/'
                                    status: Released
                                    is_nsfw: false
                                    is_paid: true
                                    has_demo: true
                                    fvn_url: 'https://fvn.li/games/chicken-police-paint-it-red'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '429':
                    $ref: '#/components/responses/RateLimited'
...
    /lists:
        get:
            tags:
                - Lists
            operationId: getLists
            summary: 'List my lists'
            x-required-ability: 'lists:read'
            parameters:
                - $ref: '#/components/parameters/Page'
            responses:
                '200':
                    description: Lists
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListPage'
                            example:
                                data:
                                    - id: 9180
                                      name: 'Currently Reading'
                                      description: null
                                      type: reading
                                      is_default: true
                                      is_public: false
                                      entries_count: 1
                                    - id: 9182
                                      name: 'Plan to Read'
                                      description: null
                                      type: plan_to_read
                                      is_default: true
                                      is_public: false
                                      entries_count: 1
                                    - id: 9181
                                      name: Completed
                                      description: null
                                      type: completed
                                      is_default: true
                                      is_public: false
                                      entries_count: 0
                                    - id: 9183
                                      name: 'On Hold'
                                      description: null
                                      type: on_hold
                                      is_default: true
                                      is_public: false
                                      entries_count: 0
                                    - id: 9184
                                      name: Dropped
                                      description: null
                                      type: dropped
                                      is_default: true
                                      is_public: false
                                      entries_count: 0
                                    - id: 9185
                                      name: 'Cozy picks for autumn'
                                      description: 'Low-stress stories to recommend to friends.'
                                      type: custom
                                      is_default: false
                                      is_public: false
                                      entries_count: 1
                                links:
                                    first: 'https://fvn.li/api/v1/lists?page=1'
                                    last: 'https://fvn.li/api/v1/lists?page=1'
                                    prev: null
                                    next: null
                                meta:
                                    current_page: 1
                                    from: 1
                                    last_page: 1
                                    links:
                                        - url: null
                                          label: '&laquo; Previous'
                                          page: null
                                          active: false
                                        - url: 'https://fvn.li/api/v1/lists?page=1'
                                          label: '1'
                                          page: 1
                                          active: true
                                        - url: null
                                          label: 'Next &raquo;'
                                          page: null
                                          active: false
                                    path: 'https://fvn.li/api/v1/lists'
                                    per_page: 20
                                    to: 6
                                    total: 6
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '429':
                    $ref: '#/components/responses/RateLimited'
        post:
            tags:
                - Lists
            operationId: postLists
            summary: 'Create a custom list'
            x-required-ability: 'lists:write'
            requestBody:
                $ref: '#/components/requestBodies/ListCreate'
            responses:
                '201':
                    description: Created
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListResponse'
                            example:
                                data:
                                    id: 9185
                                    name: 'Cozy picks'
                                    description: 'Low-stress stories to recommend to friends.'
                                    type: custom
                                    is_default: false
                                    is_public: true
                                    entries_count: 0
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '422':
                    $ref: '#/components/responses/ValidationError'
                '429':
                    $ref: '#/components/responses/RateLimited'
    '/lists/{list}':
        get:
            tags:
                - Lists
            operationId: getListsList
            summary: 'Get my list'
            x-required-ability: 'lists:read'
            parameters:
                - $ref: '#/components/parameters/List'
            responses:
                '200':
                    description: List
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListResponse'
                            example:
                                data:
                                    id: 9185
                                    name: 'Cozy picks for autumn'
                                    description: 'Low-stress stories to recommend to friends.'
                                    type: custom
                                    is_default: false
                                    is_public: false
                                    entries_count: 1
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '429':
                    $ref: '#/components/responses/RateLimited'
        patch:
            tags:
                - Lists
            operationId: patchListsList
            summary: 'Update my list'
            x-required-ability: 'lists:write'
            parameters:
                - $ref: '#/components/parameters/List'
            requestBody:
                $ref: '#/components/requestBodies/ListUpdate'
            responses:
                '200':
                    description: List
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListResponse'
                            example:
                                data:
                                    id: 9185
                                    name: 'Cozy picks for autumn'
                                    description: 'Low-stress stories to recommend to friends.'
                                    type: custom
                                    is_default: false
                                    is_public: false
                                    entries_count: 1
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '422':
                    $ref: '#/components/responses/ValidationError'
                '429':
                    $ref: '#/components/responses/RateLimited'
        delete:
            tags:
                - Lists
            operationId: deleteListsList
            summary: 'Delete my custom list'
            description: 'Default lists cannot be deleted.'
            x-required-ability: 'lists:write'
            parameters:
                - $ref: '#/components/parameters/List'
            responses:
                '204':
                    description: Deleted
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '429':
                    $ref: '#/components/responses/RateLimited'
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            description: 'User API token created on the dashboard and shown once. Tokens expire 90 days after creation; an active token can be extended on the dashboard to 90 days from extension, capped at 365 days from creation. Authenticated responses carry the expiry in X-Token-Expires-At.'
    parameters:
        Page:
            name: page
            in: query
            schema:
                type: integer
                minimum: 1
        Game:
            name: game
            in: path
            required: true
            schema:
                type: integer
            description: 'fvn.li game ID'
...
    responses:
        Unauthorized:
            description: 'Missing, invalid, or expired token'
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Message'
                    example:
                        message: Unauthenticated.
        Forbidden:
            description: 'Token lacks the required ability or action is forbidden'
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Message'
                    examples:
                        missingAbility:
                            summary: 'Token lacks the required permission'
                            value:
                                message: 'This token lacks the lists:read permission.'
                        defaultList:
                            summary: 'Default lists cannot be deleted'
                            value:
                                message: 'Default lists cannot be deleted.'
...
        ValidationError:
            description: 'Validation failed'
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ValidationError'
                    examples:
                        duplicateTag:
                            summary: 'Private tag names are unique per account, ignoring case'
                            value:
                                message: 'You already have a tag with this name.'
                                errors:
                                    name:
                                        - 'You already have a tag with this name.'
                        incompleteReorder:
                            summary: 'Reorder must list every entry exactly once'
                            value:
                                message: 'Provide every entry in this list exactly once.'
                                errors:
                                    entry_ids:
                                        - 'Provide every entry in this list exactly once.'
                        paidUpdates:
                            summary: 'Paid games cannot receive update notifications'
                            value:
                                message: 'Updates are unavailable for paid games.'
                                errors:
                                    is_receiving_updates:
                                        - 'Updates are unavailable for paid games.'
                        unknownLanguage:
                            summary: 'Languages are ISO 639-3 codes'
                            value:
                                message: 'The selected preferred_languages.0 is invalid.'
                                errors:
                                    preferred_languages.0:
                                        - 'The selected preferred_languages.0 is invalid.'
        RateLimited:
            description: 'Rate limit exceeded'
            headers:
                Retry-After:
                    description: 'Seconds until retry'
                    schema:
                        type: integer
                X-RateLimit-Limit:
                    schema:
                        type: integer
                X-RateLimit-Remaining:
                    schema:
                        type: integer
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Message'
                    example:
                        message: 'Too Many Attempts.'
    schemas:
...
        GamePage:
            allOf:
                - $ref: '#/components/schemas/PageEnvelope'
                - type: object
                  properties:
                      data:
                          type: array
                          items:
                              $ref: '#/components/schemas/Game'
...